Bring back ArtifactMap and rename file
