+++ Source/artifacts/Mappings/ComplexValueMap.ts
